Move typing utils to seprate file
